backend/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
whitenoise
dj-database-url
axios
